+++ Plugins/ProcGen/Source/NexusProcGen/Public/NProcGenDeveloperOverlayWidget.h
@@ -3,6 +3,7 @@
 
 #pragma once
 
+#include "INWidgetMessageProvider.h"
 #include "NDeveloperOverlayWidget.h"
 #include "NProcGenDeveloperOverlayWidget.generated.h"
 
@@ -11,12 +12,17 @@ class UNProcGenOperation;
 class UNListView;
 
 UCLASS(ClassGroup = "NEXUS", DisplayName = "ProcGen Developer Overlay Widget", BlueprintType, Abstract)
-class NEXUSPROCGEN_API UNProcGenDeveloperOverlayWidget : public UNDeveloperOverlayWidget
+class NEXUSPROCGEN_API UNProcGenDeveloperOverlayWidget : public UNDeveloperOverlayWidget, public INWidgetMessageProvider
 {
 	GENERATED_BODY()
 	
 	virtual void NativeConstruct() override;
 	virtual void NativeDestruct() override;
+	
+	//~INMessageProvider interface
+	virtual bool HasProvidedMessage() const override;
+	virtual FString GetProvidedMessage() const override;
+	//~End of INMessageProvider interface
 	
 	void OnOperationStatusChanged(UNProcGenOperation* Operation, const ENProcGenOperationState NewState);